+++ src/main/java/com/leo/voidminers/recipe/WeightedStack.java
@@ -5,14 +5,14 @@ import net.minecraft.world.item.ItemStack;
 
 public class WeightedStack {
     public ItemStack stack;
-    public float weight;
+    public double weight;
 
-    public WeightedStack(ItemStack stack, float weight) {
+    public WeightedStack(ItemStack stack, double weight) {
         this.stack = stack;
         this.weight = weight;
     }
 
-    public WeightedStack(Item item, float weight) {
+    public WeightedStack(Item item, double weight) {
         this.stack = item.getDefaultInstance();
         this.weight = weight;
     }
